Switch to new backend
It now uses the "replaygain" script from the rgain python module for
actually calculating gain, instead of using gstreamer. (Note that this script
calls "replaygain" as external script, because I couldn't quite figure
out how to import and use the module itself. The interfaces seemed a bit
too tightly-coupled for me to reuse.)